+++ cmd/forklift/plt/cli.go
@@ -216,6 +216,7 @@ func makeQuerySubcmds() []*cli.Command {
 			},
 		},
 		makeQueryReqSubcmds(category),
+		makeQueryImportSubcmds(category),
 		[]*cli.Command{
 			{
 				Name:     "ls-pkg",
@@ -283,6 +284,26 @@ func makeQueryReqSubcmds(category string) []*cli.Command {
 			Usage:     "Describes a repo available in the local pallet",
 			ArgsUsage: "repo_path",
 			Action:    showRepoAction,
+		},
+	}
+}
+
+func makeQueryImportSubcmds(category string) []*cli.Command {
+	return []*cli.Command{
+		{
+			Name:     "ls-imp",
+			Aliases:  []string{"list-imports"},
+			Category: category,
+			Usage:    "Lists import groups specified by the local pallet",
+			Action:   lsImpAction,
+		},
+		{
+			Name:      "show-imp",
+			Aliases:   []string{"show-import"},
+			Category:  category,
+			Usage:     "Describes an import group specified by the local pallet",
+			ArgsUsage: "import_name",
+			Action:    showImpAction,
 		},
 	}
 }