+++ internal/app/forklift/pallets-deployments.go
@@ -56,7 +56,7 @@ func ResolveDepls(
 func (d *ResolvedDepl) Check() (errs []error) {
 	if d.PkgReq.Path() != d.Def.Package {
 		errs = append(errs, errors.Errorf(
-			"required package %s does not match required package %s in deployment configuration",
+			"required package %s does not match required package %s in deployment declaration",
 			d.PkgReq.Path(), d.Def.Package,
 		))
 	}
@@ -76,7 +76,7 @@ func (d *ResolvedDepl) Check() (errs []error) {
 	return errs
 }
 
-// EnabledFeatures returns a map of the package features enabled by the deployment's configuration,
+// EnabledFeatures returns a map of the package features enabled by the deployment's declaration,
 // with feature names as the keys of the map.
 func (d *ResolvedDepl) EnabledFeatures() (enabled map[string]core.PkgFeatureSpec, err error) {
 	all := d.Pkg.Def.Features
@@ -97,7 +97,7 @@ func (d *ResolvedDepl) EnabledFeatures() (enabled map[string]core.PkgFeatureSpec
 }
 
 // DisabledFeatures returns a map of the package features not enabled by the deployment's
-// configuration, with feature names as the keys of the map.
+// declaration, with feature names as the keys of the map.
 func (d *ResolvedDepl) DisabledFeatures() map[string]core.PkgFeatureSpec {
 	all := d.Pkg.Def.Features
 	enabled := make(structures.Set[string])
@@ -544,25 +544,25 @@ func FilterDeplsForEnabled(depls []Depl) []Depl {
 }
 
 // loadDepl loads the Depl from a file path in the provided base filesystem, assuming the file path
-// is the specified name of the deployment followed by the deployment config file extension.
+// is the specified name of the deployment followed by the deployment declaration file extension.
 func loadDepl(fsys core.PathedFS, name string) (depl Depl, err error) {
 	depl.Name = name
 	if depl.Def, err = loadDeplDef(fsys, name+DeplDefFileExt); err != nil {
-		return Depl{}, errors.Wrapf(err, "couldn't load version depl config")
+		return Depl{}, errors.Wrapf(err, "couldn't load deployment declaration")
 	}
 	return depl, nil
 }
 
-// loadDepls loads all package deployment configurations from the provided base filesystem matching
+// loadDepls loads all package deployment declarations from the provided base filesystem matching
 // the specified search pattern.
-// The search pattern should not include the file extension for deployment specification files - the
+// The search pattern should not include the file extension for deployment declaration files - the
 // file extension will be appended to the search pattern by LoadDepls.
 func loadDepls(fsys core.PathedFS, searchPattern string) ([]Depl, error) {
 	searchPattern += DeplDefFileExt
 	deplDefFiles, err := doublestar.Glob(fsys, searchPattern)
 	if err != nil {
 		return nil, errors.Wrapf(
-			err, "couldn't search for package deployment configs matching %s/%s",
+			err, "couldn't search for package deployment declarations matching %s/%s",
 			fsys.Path(), searchPattern,
 		)
 	}
@@ -577,7 +577,7 @@ func loadDepls(fsys core.PathedFS, searchPattern string) ([]Depl, error) {
 		depl, err := loadDepl(fsys, deplName)
 		if err != nil {
 			return nil, errors.Wrapf(
-				err, "couldn't load package deployment config from %s", deplDefFilePath,
+				err, "couldn't load package deployment declaration from %s", deplDefFilePath,
 			)
 		}
 		depls = append(depls, depl)
@@ -600,12 +600,12 @@ func loadDeplDef(fsys core.PathedFS, filePath string) (DeplDef, error) {
 	bytes, err := fs.ReadFile(fsys, filePath)
 	if err != nil {
 		return DeplDef{}, errors.Wrapf(
-			err, "couldn't read deployment config file %s/%s", fsys.Path(), filePath,
+			err, "couldn't read deployment declaration file %s/%s", fsys.Path(), filePath,
 		)
 	}
-	config := DeplDef{}
-	if err = yaml.Unmarshal(bytes, &config); err != nil {
-		return DeplDef{}, errors.Wrap(err, "couldn't parse deployment config")
+	declaration := DeplDef{}
+	if err = yaml.Unmarshal(bytes, &declaration); err != nil {
+		return DeplDef{}, errors.Wrap(err, "couldn't parse deployment declaration")
 	}
-	return config, nil
+	return declaration, nil
 }